Make edit marker details more accessible
